+++ src/main/java/com/example/springboottpl/entity/RoleDeptBean.java
@@ -1,7 +1,6 @@
 package com.example.springboottpl.entity;
 
 import java.io.Serializable;
-import java.util.Date;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -11,15 +10,15 @@ import lombok.NoArgsConstructor;
 /**
  * 描述：角色和部门关联
  * 作者：刘飞华
- * 日期：2024-10-08 14:26:31
+ * 日期：2024-10-08 17:42:02
  */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class RoleDeptBean implements Serializable {
-  //角色ID
-  private Integer roleId; 
-  //部门ID
-  private Integer deptId; 
+    //角色ID
+    private Long roleId;
+    //部门ID
+    private Long deptId;
 }
