+++ src/main/java/com/example/springboottpl/entity/UserPostBean.java
@@ -1,7 +1,6 @@
 package com.example.springboottpl.entity;
 
 import java.io.Serializable;
-import java.util.Date;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -11,15 +10,15 @@ import lombok.NoArgsConstructor;
 /**
  * 描述：用户与岗位关联
  * 作者：刘飞华
- * 日期：2024-10-08 14:26:31
+ * 日期：2024-10-08 17:42:02
  */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserPostBean implements Serializable {
-  //用户ID
-  private Integer userId; 
-  //岗位ID
-  private Integer postId; 
+    //用户ID
+    private Long userId;
+    //岗位ID
+    private Long postId;
 }
